move fast_parse tests to test dir
