+++ openlibrary/catalog/marc/fast_parse.py
@@ -9,6 +9,9 @@ from openlibrary.catalog.utils import tidy_isbn
 re_real_book = re.compile('(pbk|hardcover|alk[^a-z]paper|cloth)', re.I)
 
 def translate(bytes_in, leader_says_marc8=False):
+    """
+    Converts MARC8 to unicode
+    """
     marc8 = MARC8ToUnicode(quiet=True)
     try:
         if leader_says_marc8:
@@ -307,7 +310,7 @@ def add_oclc(edition):
     edition.setdefault('oclc', []).append(oclc)
 
 def index_fields(data, want, check_author=True):
-    # DEPRECATED, likely/certain? to trigger exception via read_short_title
+    # DEPRECATED, has a chance of triggering exception via read_short_title
     if str(data)[6:8] != 'am': # only want books
         return None
     is_marc8 = data[9] != 'a'
@@ -507,46 +510,3 @@ def split_line(s):
             if s[m+2:marks[i+1]]:
                 ret.append(('v', s[m+2:marks[i+1]]))
     return ret
-
-def test_wrapped_lines():
-    data = open('test_data/wrapped_lines').read()
-    ret = list(handle_wrapped_lines(get_tag_lines(data, ['520'])))
-    assert len(ret) == 2
-    a, b = ret
-    assert a[0] == '520' and b[0] == '520'
-    assert len(a[1]) == 2295
-    assert len(b[1]) == 248
-
-def test_translate():
-    assert translate('Vieira, Claudio Bara\xe2una,') == u'Vieira, Claudio Bara\xfana,'
-
-def test_read_oclc():
-    from pprint import pprint
-    for f in ('oregon_27194315',):
-        data = open('test_data/' + f).read()
-        i = index_fields(data, ['001', '003', '010', '020', '035', '245'])
-        assert 'oclc' in i
-        e = read_edition(data)
-        assert 'oclc' in e
-
-def test_record():
-    assert read_edition(open('test_data/lc_0444897283').read())
-
-def test_empty():
-    assert read_edition('') == {}
-
-def bad_marc_line():
-    line = '0 \x1f\xe2aEtude objective des ph\xe2enom\xe1enes neuro-psychiques;\x1e'
-    assert list(get_all_subfields(line)) == [(u'\xe1', u'Etude objective des ph\xe9nom\xe8nes neuro-psychiques;')]
-
-def test_index_fields():
-    data = open('test_data/ithaca_college_75002321').read()
-    lccn = index_fields(data, ['010'])['lccn'][0]
-    assert lccn == '75002321'
-
-def test_ia_charset():
-    data = open('test_data/histoirereligieu05cr_meta.mrc').read()
-    line = list(get_tag_lines(data, set(['100'])))[0][1]
-    a = list(get_all_subfields(line, ia_bad_charset=True))[0][1]
-    expect = u'Cr\xe9tineau-Joly, J.'
-    assert a == expect
